pyproject.toml
[tool.poetry]
name = "jit-test"
version = "0.1.0"
description = ""
authors = ["palas1k <liserg008@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

